refactor db ops to base class
[static]
